third_party/WebKit/LayoutTests/fast/css/css-imagevalue-url.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
  test(function() {
    var cssText = 'div { background-image: url("images.jpg"); }';

    var doc = document.implementation.createHTMLDocument("");
    var styleElement = document.createElement("style");
    styleElement.textContent = cssText;
    doc.body.appendChild(styleElement);
    var cssRules = styleElement.sheet.cssRules;

    assert_equals(cssRules.length, 1);
    var compareCSSText = cssRules[0].cssText;

    assert_equals(compareCSSText, cssText);
  }, 'Testing serialization and round trip of URLs in CSS');
</script>
